# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Souqrates System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users/{userId}/ban:
    post:
      operationId: banUser
      tags:
        - admin
      summary: Ban or unban a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BanInput"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/broadcast:
    post:
      operationId: sendBroadcast
      tags:
        - admin
      summary: Send broadcast message to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastInput"
      responses:
        "200":
          description: Broadcast queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BroadcastResult"
  /admin/transactions:
    get:
      operationId: listAllTransactions
      tags:
        - admin
      summary: List all transactions (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
...
  /admin/deposits/{transactionId}/confirm:
    post:
      operationId: confirmDeposit
      tags:
        - admin
      summary: Confirm and credit a pending deposit (admin)
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit confirmation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositConfirmResult"
  /admin/deposits/{transactionId}/reject:
    post:
      operationId: rejectDeposit
      tags:
        - admin
      summary: Reject a pending deposit (admin)
      parameters:
        - name: transactionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deposit rejection result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DepositConfirmResult"
...
  /admin/subagent-applications/{applicationId}:
    patch:
      operationId: reviewSubagentApplication
      tags:
        - subagent
      summary: Approve or reject a partner application (admin)
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubagentApplicationReview"
      responses:
        "200":
          description: Application updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubagentApplicationResponse"
...
components:
  schemas:
...
    User:
      type: object
      required:
        - id
        - telegramId
        - skzBalance
        - referralCode
        - isAdmin
        - isBanned
        - totalEarned
        - createdAt
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type:
            - string
            - "null"
        firstName:
          type:
            - string
            - "null"
        lastName:
          type:
            - string
            - "null"
        photoUrl:
          type:
            - string
            - "null"
        skzBalance:
          type: number
        referralCode:
          type: string
        referredBy:
          type:
            - integer
            - "null"
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        totalEarned:
          type: number
        createdAt:
          type: string
...
    Transaction:
      type: object
      required:
        - id
        - userId
        - type
        - amount
        - status
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        type:
          type: string
          enum:
            - deposit
            - withdraw
            - transfer_in
            - transfer_out
            - commission
            - refund
        amount:
          type: number
        currency:
          type:
            - string
            - "null"
        txHash:
          type:
            - string
            - "null"
        status:
          type: string
          enum:
            - pending
            - confirmed
            - failed
        note:
          type:
            - string
            - "null"
        refUserId:
          type:
            - integer
            - "null"
        createdAt:
          type: string
    TransactionListResponse:
      type: object
      required:
        - transactions
        - total
        - page
        - limit
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    BanInput:
      type: object
      required:
        - banned
      properties:
        banned:
          type: boolean
        reason:
          type: string
    BroadcastInput:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        parseMode:
          type: string
          enum:
            - HTML
            - Markdown
        targetGroup:
          type: string
          enum:
            - all
            - active
            - referrers
    BroadcastResult:
      type: object
      required:
        - queued
        - targetCount
      properties:
        queued:
          type: boolean
        targetCount:
          type: integer
        message:
          type: string
...
    DepositConfirmResult:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - confirmed
            - already_confirmed
            - rejected
            - not_found
        transaction:
          oneOf:
            - $ref: "#/components/schemas/Transaction"
            - type: "null"
...
    SubagentApplicationResponse:
      type: object
      required:
        - id
        - telegramId
        - fullName
        - phone
        - email
        - status
        - createdAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        telegramId:
          type: string
        fullName:
          type: string
        phone:
          type: string
        email:
          type: string
        company:
          type:
            - string
            - "null"
        experience:
          type: string
        motivation:
          type: string
        status:
          type: string
        reviewNote:
          type:
            - string
            - "null"
        reviewedAt:
          type:
            - string
            - "null"
        createdAt:
          type: string
...
    SubagentApplicationReview:
      type: object
      required:
        - status
      properties:
        status:
          type: string
        reviewNote:
          type: string
...