# lib/api-zod/src/generated/types/depositConfirmResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Souqrates System API
 * OpenAPI spec version: 0.1.0
 */
import type { DepositConfirmResultStatus } from './depositConfirmResultStatus';
import type { Transaction } from './transaction';

export interface DepositConfirmResult {
  status: DepositConfirmResultStatus;
  transaction?: Transaction | null;
}